Move over util testdata into command package
