+++ agent/command/archive_zip_extract_test.go
@@ -100,7 +100,7 @@ func (s *ZipExtractSuite) TestErrorsAndNormalizedPath() {
 func (s *ZipExtractSuite) TestExtractionArchiveDoesNotExist() {
 	s.cmd.TargetDirectory = s.targetLocation
 	s.cmd.ArchivePath = filepath.Join(testutil.GetDirectoryOfFile(),
-		"testdata", "archive", "artifacts.tar.gzip")
+		"testdata", "archive", "nonexistent.tar.gz")
 
 	s.Error(s.cmd.Execute(s.ctx, s.comm, s.logger, s.conf))
 }
