+++ agent/command/testdata/archive/artifacts_in/dir1/dir2/my_symlink.txt
