+++ agent/command/testdata/archive/artifacts_in/dir1/dir2/testfile.txt
