--- agent/util/testdata/targz_me/dir1/dir2/test.pdb
@@ -1,1 +1,0 @@
-testesadhajsdhajklsdhasjkl
