--- agent/util/testdata/targz_me/dir1/dir2/testfile.txt
@@ -1,1 +1,0 @@
-test
